done with movement and placement (I believe)
done with movement and placement (I believe), but some bugs still
remain, as in images that should be there disappearing as soon as the
stop area is referred back to. I'll check on that a little more
tonight, but I think it has to do with visibility being turned off on
certain items, so I'll probably need some assistance.